Added Recipe detail page with selectable tabs and linked the current data
A lot was done on this commit to structure the recipe page's views for each tab and linking them together dynamically. At this time, everything functions as intended. The tabs do still need to be linked up to a source of information, and ways to collect that information need to be created, but as of right now this is a good base