Added add item and edit buttons
added animations and structure for adding an item to the list and editing an exisiting item
There are some visual bugs, but nothing that cannot be worked out at a later date. As this push, it does not update the values on the database, but does show a realtime update on the page.