Added Instructions base functionality
instructions now has an editor view to input text. each textbox expands based on the text entered and scrolls when the view is full. Safeguards are in place to prevent adding a new step when the previous step has not been edited. You can edit any text body at will. A  save button in the top right has been added for the user to save when done.